refactoring: authContext 기능별 파일 분리
