+++ client/src/contexts/auth/index.ts
@@ -1,0 +1,6 @@
+export { AuthProvider, useAuthContext } from './AuthContext';
+export { authActions } from './authState';
+export type { AuthState } from './authState';
+export { tokenManager } from './tokenManager';
+export { authApi } from './authApi';
+export { userApi } from './userApi';
