+++ client/src/contexts/index.tsx
@@ -1,1 +1,3 @@
-export { AuthProvider, useAuthContext } from './AuthContext';
+export { AuthProvider, useAuthContext } from './auth';
+export { tokenManager, authApi, userApi, authActions } from './auth';
+export type { AuthState } from './auth';
